requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cffi==1.15.0
click==8.1.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.15.3
Flask==2.1.1
Flask_SQLAlchemy==2.5.1
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
pip-chill==1.0.1
pip-tools==6.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
terminado==0.13.3
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
zipp==3.8.0
